app/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
python-dotenv==1.0.1
Flask-Migrate
